refactor: update user entity test folder structor
